+++ docs/changes.rst
@@ -1,6 +1,14 @@
 Changes
 =======
 
+Version 1.3.1
+-------------
+
+Unreleased
+
+- Nested blueprints inherit CSRF exemption from an exempt parent.
+  :issue:`697`
+
 Version 1.3.0
 -------------